+++ tests/testthat/test_cat_piece.r
@@ -28,10 +28,11 @@ test_that("text diagrams", {
     expect_snapshot({
         dft <- tibble(piece_side = "board_back", x=seq(1.5, 5.5, 2), y=1.5, rank=2,
                       cfg="checkers1")
-        dfb <- tibble(piece_side = "bit_back", x=1:6, y=1, suit=1:6, cfg="checkers1")
-        dfd <- tibble(piece_side = "die_face", x=1:6, y=2, suit=1:6, rank=1:6,
+        dfbb <- tibble(piece_side = "bit_back", x=1:6, y=1, suit=1:6, cfg="checkers1")
+        dfbf <- tibble(piece_side = "bit_face", x=1:6, y=2, suit=1:6, cfg="checkers1")
+        dfd <- tibble(piece_side = "die_face", x=1:6, y=3, suit=1:6, rank=1:6,
                       cfg="dice", angle = c(45, rep(0, 5)))
-        df <- dplyr::bind_rows(dft, dfb, dfd)
+        df <- dplyr::bind_rows(dft, dfbb, dfbf, dfd)
         cat_piece(df)
 
         df <- dplyr::mutate(df, cfg = gsub("checkers1", "checkers2", cfg),
@@ -91,7 +92,7 @@ test_that("text diagrams", {
                   suit=2:1, angle = c(0, 45))
     dfpf <- tibble(piece_side = "pawn_face", x=1, y=2,
                   suit=1, angle = 45)
-    dfbf <- tibble(piece_side = "bit_face", x=3, y=1, suit=3, cfg="checkers1")
+    dfbf <- tibble(piece_side = "bit_back", x=3, y=1, suit=3, cfg="checkers1")
     df <- dplyr::bind_rows(dft, dfpb, dfpf, dfbf)
     cat_piece(df)
     })